added endpoints restrictions by user role
